requirements.txt
PyQt5
Pillow
openpyxl
